// src/app/models/user.ts
export class User {
    /*  uid: string;
      email: string;
      password?: string;
      displayName: string;
      photoURL: string;
      emailVerified: boolean;*/
      name: string;
  
  
   /*   constructor() {
          this.email = '';
          this.password = '';
          this.uid = '';
          this.displayName = '';
          this.photoURL = '';
          this.emailVerified = false;
      }*/
  }
  